added matrix uniform to shader, added matrix translate and scale to draw method of renderer. Test button working with this, main fractal draw not working with this yet.

+++ MandelbrotFractal/Source/Gui.h
@@ -18,7 +18,7 @@ class Gui
 public:
 	Gui();
 	void addMouseInputEvent(int button, int action, double xpos, double ypos);
-	bool Button(int width, int height, int posX, int posY);
+	bool Button(float width, float height, float posX, float posY);
 	void resetGui();
